+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/directb2s/B2STableSettingsParser.java
@@ -299,6 +299,10 @@ public class B2STableSettingsParser extends DefaultHandler {
         }
         break;
       }
+      case "FormNoFocus": {
+        settings.setFormNoFocus(Integer.parseInt(value) == 1);
+        break;
+      }
     }
   }